+++ src/main/resources/templates/category.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <title>Przepisy według kategorii</title>
 
-  <div th:replace="fragments :: css-style"></div>
+  <div th:replace="~{fragments :: css-style}"></div>
 
 </head>
 <body>
@@ -15,7 +15,7 @@
 
   <ul style="margin-top: 20px" class="list-group">
     <li class="list-group-item" th:each="recipe: ${recipeList}">
-      <a th:href="|/potrawa/${recipe.id}|" th:text="${recipe.name}"></a>
+      <a th:href="@{|/potrawa/${recipe.id}|}" th:text="${recipe.name}"></a>
     </li>
   </ul>
