+++ src/main/resources/templates/deleteInfo.html
@@ -2,16 +2,15 @@
 <html lang="en" xmlns:th="http://www.thymeleaf.org">
 <head>
   <meta charset="UTF-8">
-  <title>Błąd</title>
+  <title>Info</title>
 
   <div th:replace="~{fragments :: css-style}"></div>
 
 </head>
 <body>
 <div class="content">
-  <h3>Coś poszło nie tak</h3>
-
-
+  <h3>Przepis usunięty</h3>
+  
   <div th:replace="fragments :: footer"></div>
 
 </div>
