--- src/main/resources/templates/recipeEdit.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Edycja przepisu</title>
-</head>
-<body>
-<div class="content">
-    <h2 th:text="|Edycja przepisu: ${recipeToEdit.name}|"></h2>
-
-    <!-- Jak dodam ten <form> to po naciśnieciu przycisku "Edytuj przepis" znajdującego się na stronie "/potrawy/id"
-     zostaje przekierowana na stronę error.html - nie wiem gdzie jest problem
-     Jak jest bez tego <form> to kieruje mnie poprawnie poprawnie na tą stronę recipeEdit.html-->
-    <form th:action="/recipe/edit" method="post" th:object="${recipeToEdit}">
-        <div class="mb-3">
-            <label for="name" class="form-label">Nazwa</label>
-            <input type="text" class="form-control" id="name" th:field="*{name}">
-        </div>
-
-
-
-    </form>
-
-    <a th:href="|/potrawa/${recipeToEdit.id}|">Powrót do przepisu</a>
-
-</div>
-
-</body>
-</html>
